Fix require deprecated module
The module `createMuiTheme` has been renamed to `createTheme` in
`material-ui` version 5, and apparently this change has been backported
to version 4. When installing @material-ui as described in the book, the
module is not found. Renamed to `createTheme` but only within the
`require`, so that the Python side remains the same.

Found on [this SO thread](https://stackoverflow.com/questions/68911647/material-ui-cant-resolve-material-ui-core-styles-createmuitheme)